:tada: Avoid using server-actions for consent and cookie tracking
