+++ aksel.nav.no/website/app/api/draft-mode/disable/route.ts
@@ -1,9 +1,9 @@
-"use server";
-
 import { draftMode } from "next/headers";
 
-export async function disableDraftModeAction() {
+export async function GET() {
   const disable = (await draftMode()).disable();
   const delay = new Promise((resolve) => setTimeout(resolve, 1000));
   await Promise.allSettled([disable, delay]);
+
+  return new Response(null, { status: 204 });
 }
